+++ src/Models/TrackAction.php
@@ -9,12 +9,10 @@ use Sfneal\Scopes\CreatedOrderScope;
 use Sfneal\Scopes\IdOrderScope;
 use Sfneal\Tracking\Builders\TrackActionBuilder;
 use Sfneal\Tracking\Models\Base\Tracking;
-use Sfneal\Tracking\Models\Traits\TrackingRelationships;
 
 class TrackAction extends Tracking
 {
     use HasFactory;
-    use TrackingRelationships;
 
     /**
      * The "booting" method of the model.