+++ src/Models/TrackActivity.php
@@ -11,13 +11,12 @@ use Sfneal\Scopes\CreatedOrderScope;
 use Sfneal\Scopes\IdOrderScope;
 use Sfneal\Tracking\Builders\TrackActivityBuilder;
 use Sfneal\Tracking\Models\Base\Tracking;
-use Sfneal\Tracking\Models\Traits\TrackingRelationships;
+use Sfneal\Tracking\Utils\ModelAdapter;
 
 class TrackActivity extends Tracking
 {
     // todo: add use of polymorphic relationships
     use HasFactory;
-    use TrackingRelationships;
 
     /**
      * The "booting" method of the model.
@@ -89,11 +88,13 @@ class TrackActivity extends Tracking
     /**
      * Related TrackTraffic data.
      *
+     * // todo: add tests
+     *
      * @return BelongsTo
      */
     public function tracking()
     {
-        return $this->belongsTo(TrackTraffic::class, 'request_token', 'request_token')
+        return $this->belongsTo(ModelAdapter::TrackTraffic(), 'request_token', 'request_token')
             ->withoutGlobalScope(SoftDeletingScope::class);
     }