+++ CHANGELOG.md
@@ -69,3 +69,10 @@ All notable changes to `tracking` will be documented in this file
 
 ## 0.5.2 - 2021-04-16
 - fix issue with `Response` type hinting as `RedirectResponse` $response params are also acceptable
+
+
+## 0.6.0 - 2021-04-16
+- make `ModelAdapter` for dynamically retrieving model classes
+- add 'models' section to 'tracking' config that allows for overwriting default models with custom extensions
+- add use of `ModelAdapter` for accessing `TrackAction`, `TrackActivity` & `TrackTraffic` models
+- cut `TrackingRelationship` trait and use in models